+++ symengine/series.h
@@ -20,7 +20,7 @@ public:
     virtual RCP<const Basic> as_basic() const = 0;
     virtual umap_int_basic as_dict() const = 0;
     virtual RCP<const Basic> get_coeff(int) const = 0;
-    virtual long get_degree() const = 0;
+    virtual unsigned get_degree() const = 0;
     virtual const std::string &get_var() const = 0;
 };
 
@@ -30,14 +30,14 @@ class SeriesBase : public SeriesCoeffInterface
 protected:
     const Poly p_;
     const std::string var_;
-    const int degree_;
+    const unsigned degree_;
 
 public:
-    inline SeriesBase(Poly p, std::string var, int degree)
+    inline SeriesBase(Poly p, std::string var, unsigned degree)
         : p_(std::move(p)), var_(var), degree_(degree)
     {
     }
-    inline virtual long get_degree() const
+    inline virtual unsigned get_degree() const
     {
         return degree_;
     }
@@ -100,9 +100,7 @@ public:
             }
             return make_rcp<Series>(Poly(p_ + o.p_), var_, deg);
         } else if (other.get_type_code() < Series::type_code_id) {
-            Poly p = Series::series(other.rcp_from_this(), var_,
-                                    static_cast<unsigned>(degree_))
-                         ->p_;
+            Poly p = Series::series(other.rcp_from_this(), var_, degree_)->p_;
             return make_rcp<Series>(Poly(p_ + p), var_, degree_);
         } else {
             return other.add(*this);
@@ -118,15 +116,10 @@ public:
                 throw NotImplementedError(
                     "Multivariate Series not implemented");
             }
-            return make_rcp<Series>(
-                Series::mul(p_, o.p_, static_cast<unsigned>(deg)), var_, deg);
+            return make_rcp<Series>(Series::mul(p_, o.p_, deg), var_, deg);
         } else if (other.get_type_code() < Series::type_code_id) {
-            Poly p = Series::series(other.rcp_from_this(), var_,
-                                    static_cast<unsigned>(degree_))
-                         ->p_;
-            return make_rcp<Series>(
-                Series::mul(p_, p, static_cast<unsigned>(degree_)), var_,
-                degree_);
+            Poly p = Series::series(other.rcp_from_this(), var_, degree_)->p_;
+            return make_rcp<Series>(Series::mul(p_, p, degree_), var_, degree_);
         } else {
             return other.mul(*this);
         }
@@ -149,40 +142,33 @@ public:
                 p = Series::pow(
                     p_, (static_cast<unsigned>(
                             down_cast<const Integer &>(other).neg()->as_int())),
-                    static_cast<unsigned>(deg));
-                p = Series::series_invert(p, Series::var(var_),
-                                          static_cast<unsigned>(deg));
+                    deg);
+                p = Series::series_invert(p, Series::var(var_), deg);
                 return make_rcp<Series>(p, var_, deg);
             }
             p = Series::pow(p_,
                             static_cast<unsigned>(
                                 (down_cast<const Integer &>(other).as_int())),
-                            static_cast<unsigned>(deg));
+                            deg);
             return make_rcp<Series>(p, var_, deg);
         } else if (other.get_type_code() < Series::type_code_id) {
-            p = Series::series(other.rcp_from_this(), var_,
-                               static_cast<unsigned>(degree_))
-                    ->p_;
+            p = Series::series(other.rcp_from_this(), var_, degree_)->p_;
         } else {
             return other.rpow(*this);
         }
         p = Series::series_exp(
-            Poly(p * Series::series_log(p_, Series::var(var_),
-                                        static_cast<unsigned>(deg))),
-            Series::var(var_), static_cast<unsigned>(deg));
+            Poly(p * Series::series_log(p_, Series::var(var_), deg)),
+            Series::var(var_), deg);
         return make_rcp<Series>(p, var_, deg);
     }
 
     virtual RCP<const Number> rpow(const Number &other) const
     {
         if (other.get_type_code() < Series::type_code_id) {
-            Poly p = Series::series(other.rcp_from_this(), var_,
-                                    static_cast<unsigned>(degree_))
-                         ->p_;
+            Poly p = Series::series(other.rcp_from_this(), var_, degree_)->p_;
             p = Series::series_exp(
-                Poly(p_ * Series::series_log(p, Series::var(var_),
-                                             static_cast<unsigned>(degree_))),
-                Series::var(var_), static_cast<unsigned>(degree_));
+                Poly(p_ * Series::series_log(p, Series::var(var_), degree_)),
+                Series::var(var_), degree_);
             return make_rcp<Series>(p, var_, degree_);
         } else {
             throw SymEngineException("Unknown type");
